refactor code to display Salesforce Objects, simplify license check filter in POM
refactor code to display Salesforce Objects and simplify license check filter in POM.
