+++ src/main/java/com/salesforce/dataloader/ui/DataSelectionPage.java
@@ -93,40 +93,7 @@ public class DataSelectionPage extends LoadPage {
 
         Composite comp = new Composite(parent, SWT.NONE);
         comp.setLayout(gridLayout);
-
-        Label label = new Label(comp, SWT.RIGHT);
-        label.setText(Labels.getString("DataSelectionPage.selectObject")); //$NON-NLS-1$
-        GridData data = new GridData();
-        label.setLayoutData(data);
-
-        // Add a checkbox to toggle filter
-        Button filterAll = new Button(comp, SWT.CHECK);
-        filterAll.setText(Labels.getString("DataSelectionPage.showAll")); //$NON-NLS-1$
-        data = new GridData();
-        filterAll.setLayoutData(data);
-        
-        Text search = new Text(comp, SWT.SEARCH | SWT.ICON_CANCEL | SWT.ICON_SEARCH);
-        data = new GridData(GridData.FILL, GridData.FILL, true, true);
-        data.widthHint = 140;
-        search.setLayoutData(data);
-        search.addSelectionListener(new SelectionAdapter() {
-            public void widgetDefaultSelected(SelectionEvent e) {
-                lv.refresh();
-            }
-        });
-        
-        this.filter = new EntityFilter(search, filterAll);
-        lv = new ListViewer(comp, SWT.H_SCROLL | SWT.V_SCROLL | SWT.BORDER);
-        lv.setContentProvider(new EntityContentProvider());
-        lv.setLabelProvider(new EntityLabelProvider());
-        lv.setInput(null);
-        data = new GridData(GridData.FILL, GridData.FILL, true, true);
-        data.heightHint = 140;
-        data.widthHint = 140;
-        lv.getControl().setLayoutData(data);
-        lv.addFilter(filter);
-        lv.setSorter(new EntityViewerSorter());
-
+        lv = EntitySelectionListViewerUtil.getEntitySelectionListViewer(comp);
         lv.addSelectionChangedListener(new ISelectionChangedListener() {
             @Override
             public void selectionChanged(SelectionChangedEvent event) {
@@ -139,13 +106,6 @@ public class DataSelectionPage extends LoadPage {
         if (controller.isLoggedIn()) {
             setInput(controller.getEntityDescribes());
         }
-
-        filterAll.addSelectionListener(new SelectionAdapter() {
-            @Override
-            public void widgetSelected(SelectionEvent event) {
-                lv.refresh();
-            }
-        });
 
         new Label(comp, SWT.NONE);
 
@@ -164,7 +124,7 @@ public class DataSelectionPage extends LoadPage {
         //now select the csv
 
         Composite compChooser = new Composite(comp, SWT.NONE);
-        data = new GridData(GridData.FILL_HORIZONTAL | GridData.VERTICAL_ALIGN_END);
+        GridData data = new GridData(GridData.FILL_HORIZONTAL | GridData.VERTICAL_ALIGN_END);
         data.widthHint = 400;
         compChooser.setLayoutData(data);
